[Textbased] Fetch correctly the types of the query fields (#162982)
## Summary

This PR is solving the following wrong behavior in text based languages.

When we are trying to depict a SQL/ESQL suggestion in Discover it is
necessary to know the type of the field. What we did so far was sending
from Discover to Lens only the column names and we were trying in Lens
to fetch the type from the dataview fields. But this doesnt cover all
the cases as both SQL (and ESQL) can create fields which cant be found
in a dataview. For example:

- Use of as `SELECT bytes, geo.dest as random_name FROM
"kibana_sample_data_logs"` in SQL
- Use rename (in ESQL)
- Create an aggregation, this will create a new variable in both
languages
- Using eval in ESQL
- ....

This PR solves the problem by passing the entire datatableColumn in the
Lens api and not only the name of the column. The types of the fields
are returned by sql, esql apis so we already have the information. We
just need to pass it in Lens.

**Fetching the correct types is very important as better suggestions
will be suggested from Lens.**

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>