Increase server close timeout
Prior to Node.js 19.0.0, calling close while an idle client was connected to a
server would not actually close the server before the client disconnected. This
behavior was modified in Node.js 19.0.0:

https://github.com/nodejs/node/pull/43522

This resulted in our mock server being closed before we expected and as a
result, some of the tests in `cluster.test.js` would fail. This is because the
1 second timeout is lower than the exponential backoff logic in the
`_autoRetry` function inside the `native_realm.js` file.

To get around this issue, the timeout have simply been incresed to 5 seconds
instead of 1. This is not pretty, but it gets the job done.