[FTR] unskip tsvb time series tests for Chrome (#163510)
## Summary

Related to #162995

This PR unskip TSVB tests for Chrome browser since it is proved to be
stable

https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/2836
100x passed

https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/2835
100x passed
On Firefox the flakiness is related to Terms 2nd aggregation field
sometimes is not selected. I tested it manually in Firefox 116 and was
able to set fields, though I have a feeling that values are not always
selected on click in the drop-down. But I didn't see any errors in
console.

I also returned back retry for dropdown selection I removed in #161202
though flaky-test-runner proves there is no need. Let's have just to
keep logic as before my PR.