[ResponseOps] use Data Streams for AAD indices in serverless
resolves https://github.com/elastic/kibana/issues/154266

Changes the way the alerts-as-data (AAD) indices are created and
written to, to allow them to be built as they have been in the past
(alias and backing indices created manually) OR as an ES Data Stream.

Serverless will use Data Streams, other environments will use the
existing alias and backing indices.  The determination is made by
optionally including the `serverless` plugin, and determining if it's
available.