[Response Ops][Task Manager] Expose SLI metrics in HTTP API - Take 2 (#163652)
## Summary

Redo of this PR https://github.com/elastic/kibana/pull/162178 but
without the `native-hdr-histogram` library which caused issues in the
serverless build. In the future we may want to pursue generating a
custom build of this native library but for our current purposes, a
simple bucketed histogram should suffice. The only changes from the
original PR are in this commit:
https://github.com/elastic/kibana/pull/163652/commits/dde5245deddf1470f548181c22538cb819fd089e,
where we create a `SimpleHistogram` class to bucket task claim durations
into `100ms` buckets.

Please reference the original PR for more description about this HTTP
API

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>