[Security Solution][Skipped Test] Saved query validation error test (#163050)
## Summary

Original ticket: https://github.com/elastic/kibana/issues/159060 

This PR un-skips test which was disabled after the Rule Editing page
[refactoring](https://github.com/elastic/kibana/pull/157749). There we
stopped fields validation on page loading. To be able to show the
"failed to load saved query" error on page loading we force the field
validation when we failed to load a saved query.