[QA] Change "contains" text for Dev Console CCS test (#163814)
## Summary

This is a fix for the flaky test
https://github.com/elastic/kibana/issues/163365 that checks Dev Console
with CCS. Before we were looking for the response to the query to
contain `extension: jpg` which might or might not show up in the
viewport based on the logstash data ingested (some of the documents have
a multi-line `message` field that will scroll everything else out of the
viewport).
<img width="1210" alt="test"
src="https://github.com/elastic/kibana/assets/5196019/729f4645-db48-46bc-8376-aee8602ef500">

With the change we're looking for the `_index` metafield which can also
confirm that the search worked and it returned results from a remote
cluster. There shouldn't be any instance of it not showing up in the
viewport since it's one of the first fields in a search hit.