[7.17] Test cypress OOM and chrome upgrade (#163363)
This PR introduces several changes to the Cypress config for 7.17 that
tries to mimic the state we currently have on main. The most notable
changes are the upgrade to cypress v12 and the addition of the
parallelisation. Both have required fundamentally bigger changes to what
we had before on that branch.

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>
Co-authored-by: Tiago Costa <tiago.costa@elastic.co>
Co-authored-by: Patryk Kopycinski <contact@patrykkopycinski.com>