[Response Ops][RAM] Add RBAC Support for Multi-Consumer Rule Creation in O11Y and Stack Management  (#162605)
## Summary

Resolves: 
- https://github.com/elastic/kibana/issues/162484
- https://github.com/elastic/kibana/issues/160677

This PR adds the ability for `logs` and/or `infrastructure` only users
to create and modify ES Query and new Generic Threshold rules. The
`ensureAuthorized` function is modified and simplified to support this
use case, by skipping producer authorization and only authorizing for
consumers. When the consumer is `alerts`, we will consider this legacy
and replace it with the rule’s producer (`consumer = ruleType.producer`)

There is now a dropdown in the rule form to prompt the user when
creating ES Query/Generic threshold rules to select the consumer based
on their authorized consumers (we can no longer use `alerts` for these).
If there is only 1 option, then the dropdown will not be shown and the
option will be chosen automatically.

Generic threshold rules will have the following possible consumers:
 - slo
 - infrastructure
 - logs
 - apm
 - uptime 

ES query rules will have the following possible consumers:
 - slo
 - infrastructure
 - logs
 - apm
 - uptime 
 - stackAlerts

## To Test:
### Single Consumer:
1. Create a user with only `logs` feature enabled (ensuring
`stackAlerts` is not enabled).
2. Navigate to the O11Y rule management page
3. Click the create rule button
4. Assert that both ES query and generic threshold rules are available
5. Click ES query and fill out the relevant information and create the
rule
6. Assert that the rule created has `logs` set in the `consumer` field
7. Repeat 5-6 for the generic threshold rule
8. Repeat 2-7 but on the Stack Management rules page  
9. Repeat 1-8 for the `infrastructure` feature. 

### Multiple Consumers:
1. Create a user with `logs`, `infrastructure` and `apm` features
enabled (ensuring `stackAlerts` is not enabled).
2. Navigate to the O11Y rule management page
3. Click the create rule button
4. Assert that both ES query and generic threshold rules are available
5. Click ES query and fill out the relevant information and create the
rule
6. A dropdown should prompt the user to select between 1 of the 3
consumers, select 1
7. Assert that the rule was created with the selected consumer
8. Repeat 5-7 for the generic threshold rule
9. Repeat 2-8 but on the Stack Management rules page

![Screenshot from 2023-08-08
16-45-43](https://github.com/elastic/kibana/assets/74562234/8c5b644a-8bab-4c1b-93b0-acfa956af19c)

![consumer_dropdown_open](https://github.com/elastic/kibana/assets/74562234/a03b7e97-e90e-4bbc-bed0-94a6c677d31d)


### Checklist
- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>