fix(NA): source classifier to indentify xml as static files (#164109)
This PR fixes a bug in the source classifier as it is currently not
identifying xml as a static file.

As part of this I've also identified a TODO in the source classifier
related with the evaluating order.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit a17e307dc8840eadfc97ae3d5cf8d68052d5cd15)