Failing test: Security Solution Cypress.x-pack/test/security_solution_cypress/cypress/e2e/detection_response/rule_creation/custom_query_rule·cy·ts - Custom query rules Custom detection rules deletion and edition Deletion Deletes one rule from detail page Deletes one rule from detail page #163977 (#164327)
## Summary

Fixes: https://github.com/elastic/kibana/issues/163977
Fixes: https://github.com/elastic/kibana/issues/163568

These changes fix the issue with actions popover and the way we check
whether it is closed on rule's detail page.

The issue happens due to the fact that after we closed the popover we
should be testing `should('not.exist')` instead of
`should('not.be.visible')`.

(cherry picked from commit 49f95280518d9c3c75a6efd1f2ca488fac18eaf4)