[Profiling] checking viewer resources in the admin check (#164086)
We identified a bug in the set up check where for admin users we should
also include the checks that were made in the step above.

We also decided to revert the changes in the package names.

---------

Co-authored-by: Francesco Gualazzi <inge4pres@users.noreply.github.com>
(cherry picked from commit 2e854d65ce0bc7bbe579d90582151f2ebe50d004)