[ci] Fix on-merge labels
Currently some on merge steps fetch labels from the merged pull request
by parsing the first line of the commit message.  This convention
usually works, but it overlaps with the convention of skipping test
suites linking to an issue instead of a pull request.  Errors
are thrown when a step attempts to pull metadata from an unset value.

This sets a default value of an empty string if the label metadata is
unset and cleans up the log message when an error occurs.