[ML] Fix anomaly detection module manifest queries for kibana sample data sets (#164332)
Follow up to https://github.com/elastic/kibana/pull/119635
All modules which contain a query in their manifest file should have a
filter to avoid querying cold and frozen tiers.
The original PR did not include the sample data sets as it was assumed
it was not necessary due to these data sets never being added to cold or
frozen tiers.
However it was overlooked that these queries will be run on any index
pattern passed to the `/internal/ml/modules/recognize` endpoint and so
has the potential to negatively impact the speed of all calls to this
endpoint, especially if the index pattern contains multiple indices and
wildcards.

(cherry picked from commit cd219942d12083897d7c56723f83da2a2f3c2586)